changed point id to uint32_t, implemented custom tour segment insertion

+++ k-opt/fileio/Tour.h
@@ -2,6 +2,7 @@
 
 // This represents a specific TSP solution.
 
+#include <cstdint>
 #include <fstream>
 #include <iostream>
 #include <string>
@@ -14,11 +15,11 @@ class Tour
 public:
     Tour() = default;
     Tour(const char* file_path);
-    int count() const { return m_point_ids.size(); }
-    const std::vector<int>& point_ids() const { return m_point_ids; }
+    size_t count() const { return m_point_ids.size(); }
+    const std::vector<uint32_t>& point_ids() const { return m_point_ids; }
 
 private:
-    std::vector<int> m_point_ids;
+    std::vector<uint32_t> m_point_ids;
 };
 
 } // namespace fileio
